+++ src/internal_modules/roc_audio/target_speexdsp/roc_audio/resampler_speex.h
@@ -59,7 +59,7 @@ public:
     virtual size_t pop_output(Frame& out);
 
     //! How many samples were pushed but not processed yet.
-    virtual size_t n_left_to_process() const;
+    virtual float n_left_to_process() const;
 
 private:
     void report_stats_();
